restructure directories to match proposed standards
