+++ .gitignore
@@ -1,9 +1,14 @@
 *.ipynb_checkpoints
+*.img*
+*.envi*
+*.hdr*
 *.nc*
 /data/example_out.csv
 /data/isla_gaviota_3.geojson
 /data/outputs
 /data/envi
+/python/daac_data_download_python
+/python/emit_utils
 /ignore
 *__pycache__*
 /data/emit_asset_urls.txt
